requirements.txt
ultralytics
streamlit
opencv-python-headless
gTTS
Pillow
numpy